requirements.txt
python-telegram-bot[job-queue]>=20.0
msal>=1.24.0
requests>=2.31.0
youtube-transcript-api>=0.6.0
feedparser>=6.0.0
fpdf2>=2.7.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
